HYD-13 refactor: search-result api 파일 삭제 -> dataset api 파일로 통합 & dataset 상세정보 api 추가
